% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_historicals.R
\name{rh_historicals}
\alias{rh_historicals}
\title{Get the historical quotes for a symbol.}
\usage{
rh_historicals(symbols, interval = "day", span = "year",
  bounds = "regular", keep_meta = FALSE)
}
\arguments{
\item{interval}{The interval: week|day|10minute|5minute|null(all)}

\item{span}{day|week|year|5year|all}

\item{bounds}{extended|regular|trading}

\item{symbol}{The shorthand symbols.}
}
\description{
Get the historical quotes for a symbol.
}
\examples{

library(robinhoodr)

symbol <- c("MSFT")

rh_historicals(symbol = symbol) # Get the last years closing prices.
rh_historicals(symbol = symbol, interval = "5minute", span = "day", bounds = "trading")
rh_historicals(symbol = symbol, interval = "5minute", span = "day", bounds = "extended")
rh_historicals(symbol = symbol, interval = "5minute", span = "day", bounds = "regular")
}

