% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_instruments.R
\name{rh_instruments}
\alias{rh_instruments}
\title{Get a big list of all the instruments.}
\usage{
rh_instruments()
}
\description{
Get a big list of all the instruments.
}
\examples{

robin_inst <- rh_instruments()

}

