% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_instruments_info.R
\name{rh_instruments_info}
\alias{rh_instruments_info}
\title{Function to get the fundamentals for the symbols.}
\usage{
rh_instruments_info(symbols)
}
\arguments{
\item{symbols}{Symbols}
}
\description{
Function to get the fundamentals for the symbols.
}
\examples{

library(robinhood)

rh_instruments_info(symbols = c("MSFT", "FB", "GOOG"))

}

