% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_instruments_search.R
\name{rh_instruments_search}
\alias{rh_instruments_search}
\title{Instrument keyword search.}
\usage{
rh_instruments_search(keyword)
}
\arguments{
\item{keyword}{Keyword to search for.}
}
\description{
This function lets you search for a keyword
that fx might be found in the name.
}
\examples{

library(robinhood)

rh_instrument_search(keyword = "oil")
}

