% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_instruments_split.R
\name{rh_instruments_split}
\alias{rh_instruments_split}
\title{Function to get the split info on a symbol.}
\usage{
rh_instruments_split(symbols)
}
\arguments{
\item{symbols}{Symbols.}
}
\description{
Function to get the split info on a symbol.
}
\examples{

library(robinhood)

rh_instruments_split(c("MSFT", "JDST", "FB"))
}

