% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_market_hours.R
\name{rh_market_hours}
\alias{rh_market_hours}
\title{Get the market info for a specific market.}
\usage{
rh_market_hours(mic, date = Sys.Date())
}
\arguments{
\item{mic}{Market identifier code.}
}
\description{
Get the market info for a specific market.
}
\examples{

robin_mkt_hours <- rh_market_hours(mic = "BATS")

}

