% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_market_info.R
\name{rh_market_info}
\alias{rh_market_info}
\title{Get the market info for a specific market.}
\usage{
rh_market_info(mic)
}
\arguments{
\item{mic}{Market identifier code.}
}
\description{
Get the market info for a specific market.
}
\examples{

robin_mkt_info <- rh_market_info(mic = "BATS")
}

