% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_quote.R
\name{rh_quote}
\alias{rh_quote}
\title{Get a quote for one or more symbols.}
\usage{
rh_quote(symbols)
}
\arguments{
\item{symbols}{The shorthand symbols.}
}
\description{
Get a quote for one or more symbols.
}
\examples{

library(robinhoodr)

symbols <- c("MSFT", "FB", "TSLA")

rh_quote(symbols = symbols)
}

