% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_historicals.R
\name{rh_historicals}
\alias{rh_historicals}
\title{Get the historical quotes for a symbol.}
\usage{
rh_historicals(symbols, interval = "day", span = "year",
  bounds = "regular", keep_meta = FALSE, to_xts = TRUE)
}
\arguments{
\item{interval}{The interval: week|day|10minute|5minute}

\item{span}{The span: day|week|year}

\item{bounds}{extended|regular|trading}

\item{symbol}{The shorthand symbols.}
}
\description{
Get the historical quotes for a symbol.
}
\examples{

library(robinhoodr)

symbol <- c("MSFT")

rh_historicals(symbol = symbol) # Get the last years closing prices.
rh_historicals(symbol = symbol, interval = "5minute", span = "day", bounds = "trading")
rh_historicals(symbol = symbol, interval = "5minute", span = "week", bounds = "regular")
rh_historicals(symbol = symbol, interval = "10minute", span = "day", bounds = "regular")
rh_historicals(symbol = symbol, span = "year", keep_meta = TRUE, to_xts = FALSE)
rh_historicals(symbol = symbol, span = "", to_xts = FALSE)
rh_historicals(symbol = symbol, to_xts = TRUE)
}

